fix: Fixed mlx inputs and reservior, Fixed ambient light with patchwork
Inputs are checked at every render frame with there was an update from
key press or mouse movement. Sets a flag to cleanse reservoir if
movement.

Ambient light is added with 2 ways, when light hits darkness and all
objects have emission that is iteratively added to the reservoir.
Warning, it is not fully tested.